pyproject.toml
[project]
name = "esthetician-software"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "google-api-python-client>=2.184.0",
    "google-auth>=2.41.1",
    "pydantic>=2.12.2",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
